requirements.txt
Flask==2.2.3
flask-restful==0.3.10
psycopg2==2.9.6
python-dotenv==1.0.0